Main-Application-main/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.6
django-ckeditor==6.1.0
django-email-verification==0.1.0
django-extensions==3.1.3
django-js-asset==1.2.2
django-recaptcha==2.0.6
idna==3.2
oauthlib==3.1.1
# Pillow==8.3.1
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
pymysql
pymupdf==1.19.3
pandas
django-resized
django-tinymce
django-taggit